fix(timers): Groups

+++ src/groups/timer.lua
@@ -5,10 +5,12 @@ timers:create(10000, function()
             if player:IsFakeClient() == 0 then
                 local expiretime = expiretimes[tostring(player:GetSteamID())]
                 local group = player:vars():Get("vip.group")
-                if (expiretime or os.time()) - os.time() < 0 or not groupsMap[group] then
-                    db:Query(string.format("delete from %s where steamid = '%s' limit 1", config:Fetch("vips.table_name"), tostring(player:GetSteamID())))
-                    player:vars():Set("vip.group", "none")
-                    expiretimes[tostring(player:GetSteamID())] = nil
+                if group ~= "none" and group ~= 0 then
+                    if (expiretime or os.time()) - os.time() < 0 or not groupsMap[group] then
+                        db:Query(string.format("delete from %s where steamid = '%s' limit 1", config:Fetch("vips.table_name"), tostring(player:GetSteamID())))
+                        player:vars():Set("vip.group", "none")
+                        expiretimes[tostring(player:GetSteamID())] = nil
+                    end
                 end
             end
         end
